Modularized js
Minor tweaks, and modularization of code
