+++ src/scss/custom/_header.scss
@@ -1,7 +1,5 @@
 header {
-  background-color: black;
   div {
-    text-decoration: none !important;
     img {
       width: 40px;
       height: 56px;
